+++ app/src/main/kotlin/com/silvertown/android/dailyphrase/DailyPhraseApplication.kt
@@ -1,0 +1,13 @@
+package com.silvertown.android.dailyphrase
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+import timber.log.Timber
+
+@HiltAndroidApp
+class DailyPhraseApplication : Application() {
+    override fun onCreate() {
+        super.onCreate()
+        Timber.plant(Timber.DebugTree())
+    }
+}
