--- app/src/main/kotlin/com/slivertown/android/dailyphrase/DailyPhraseApplication.kt
@@ -1,13 +1,0 @@
-package com.slivertown.android.dailyphrase
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-import timber.log.Timber
-
-@HiltAndroidApp
-class DailyPhraseApplication : Application() {
-    override fun onCreate() {
-        super.onCreate()
-        Timber.plant(Timber.DebugTree())
-    }
-}
