+++ data/src/androidTest/java/com/silvertown/android/dailyphrase/data/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.silvertown.android.dailyphrase.domain
+package com.silvertown.android.dailyphrase.data
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.silvertown.android.dailyphrase.domain.test", appContext.packageName)
+        assertEquals("com.silvertown.android.dailyphrase.data.test", appContext.packageName)
     }
 }
