--- presentation/src/main/kotlin/com/slivertown/android/dailyphrase/presentation/ui/MainActivity.kt
@@ -1,15 +1,0 @@
-package com.slivertown.android.dailyphrase.presentation.ui
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
-import com.slivertown.android.dailyphrase.presentation.R
-
-class MainActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        installSplashScreen()
-
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-    }
-}
